enable DEMO mode

+++ src/com/bpshparis/hak/ccadalogger/Position.java
@@ -2,20 +2,20 @@ package com.bpshparis.hak.ccadalogger;
 
 public class Position {
 
-	long lat = 0L;
-	long lng = 0L;
+	double lat = 0D;
+	double lng = 0D;
 	long timestamp = 0L;
 
-	public long getLat() {
+	public double getLat() {
 		return lat;
 	}
-	public void setLat(long lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
-	public long getLng() {
+	public double getLng() {
 		return lng;
 	}
-	public void setLng(long lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
 	public long getTimestamp() {
